handlers: refactor types into versioned types
Moved the v1 response objects into a package of their own,
under types/v1
